move safe logging to internal
