+++ resource-managers/kubernetes/core/pom.xml
@@ -90,12 +90,6 @@
     </dependency>
 
     <dependency>
-      <groupId>com.palantir.safe-logging</groupId>
-      <artifactId>safe-logging</artifactId>
-      <version>1.5.1</version>
-    </dependency>
-
-    <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-core</artifactId>
       <scope>test</scope>
